changed organization and added test functions
