+++ gameLoop.py
@@ -1,5 +1,5 @@
 from GameVar import GameVar
-from Plays import Play, KickOff, FieldGoal, RushPlay, PassPlay, Punt
+from Plays import KickOff, FieldGoal, RushPlay, PassPlay, Punt
 from Game import Game
 from Team import Team
 import random
@@ -161,3 +161,5 @@ def runSimulation(homeTeam = 'BUF', awayTeam = 'KC', numGames = 10): #This will 
         game_scores.append((a, b))
         games.append(c)
     return games, game_scores #Returns a list of game dataframes, and a list of tuples containing game scores
+
+runSimulation()
